Moved attributionHandler, minor fixes
